rhythm: Switch from double to float
Why waste spaces?

+++ src/main/java/cse213/badc/rhythm/Analytics.java
@@ -26,6 +26,6 @@ public abstract class Analytics implements Serializable {
         return lastUpdated;
     }
 
-    public abstract double getPerformanceScore();
+    public abstract float getPerformanceScore();
     public abstract String getStatus();
 }
