+++ src/main/java/cse213/badc/rhythm/AnalyticsSummary.java
@@ -3,7 +3,7 @@ package cse213.badc.rhythm;
 import java.util.ArrayList;
 
 public interface AnalyticsSummary {
-    double getAveragePerformance();
+    float getAveragePerformance();
     int getTotalCount();
     int getApprovedCount();
     String getTopPerformingRegion();
